fringematrix5-yvw: extract request timeout middleware to server/middleware/timeout.ts
Move the inline timeout middleware and REQUEST_TIMEOUT_MS constant from
server.ts into server/middleware/timeout.ts. Update the test file to
import the real implementation instead of reimplementing it verbatim,
eliminating the drift risk.
