+++ server/server.ts
@@ -7,6 +7,7 @@ import { list, ListBlobResultBlob } from '@vercel/blob';
 import { fileURLToPath } from 'url';
 import { VALID_CONTENT_PAGES } from '../shared/types.js';
 import type { ContentPage } from '../shared/types.js';
+import { timeoutMiddleware } from './middleware/timeout.js';
 
 interface Campaign {
   id: string;
@@ -79,29 +80,7 @@ const app: Application = express();
 app.disable('x-powered-by');
 const PORT = process.env['PORT'] || 3000;
 
-// Defense-in-depth: send a 503 if an async request handler takes longer than
-// 30 seconds. This handles cases such as an upstream dependency hanging or a
-// very slow network call. Note: a synchronous CPU-blocking operation (e.g. a
-// ReDoS in route matching) cannot be interrupted by a setTimeout callback;
-// protection against that class of attack comes from using a patched
-// path-to-regexp (via Express 5), not from this middleware.
-const REQUEST_TIMEOUT_MS = 30_000;
-app.use((_req: Request, res: Response, next) => {
-  let timedOut = false;
-  const timer = setTimeout(() => {
-    timedOut = true;
-    if (!res.headersSent) {
-      res.status(503).json({ error: 'Request timeout' });
-    }
-  }, REQUEST_TIMEOUT_MS);
-  // Store the flag on res.locals so downstream handlers can skip writing
-  // their response after the timeout has already fired.
-  res.locals['timedOut'] = () => timedOut;
-  // Clear the timer when the response completes so it doesn't fire late.
-  res.on('finish', () => clearTimeout(timer));
-  res.on('close', () => clearTimeout(timer));
-  next();
-});
+app.use(timeoutMiddleware);
 
 // Project root is one level up from this file (which lives in server/)
 const PROJECT_ROOT = path.join(__dirname, '..');
